django_test/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
#testing
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.3.0
pytest-selenium==2.0.1